requirements.txt
pandas==2.0.3 
numpy==1.24.3 
beautifulsoup4==4.12.2 
requests==2.31.0 
playwright==1.39.0 
mysql-connector-python==8.1.0 
sqlalchemy==2.0.23 
scikit-learn==1.3.0 
xgboost==2.0.2 
joblib==1.3.2 
streamlit==1.28.1 
plotly==5.17.0 
matplotlib==3.7.2 
schedule==1.2.0 
python-dotenv==1.0.0 
